requirements.txt
# Requisitos principales para procesamiento de imágenes satelitales VNP46A1
# Python >= 3.8

# Procesamiento de datos
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0
h5py>=3.6.0,<4.0.0
scipy>=1.7.0,<2.0.0
pyarrow>=10.0.0

# Visualización
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0
pillow>=8.0.0

# Análisis estadístico y series temporales
statsmodels>=0.13.0

# Geometría geoespacial
geopy>=2.2.0
shapely>=1.8.0
geopandas>=0.10.0
pyproj>=3.0.0

# Validación de datos
pydantic>=2.0.0
python-dotenv>=1.0.0

# Operaciones asíncronas
aiohttp>=3.8.0,<4.0.0
nest_asyncio>=1.5.0

# Web scraping
beautifulsoup4>=4.10.0
requests>=2.26.0

# Dependencias comunes
python-dateutil>=2.8.0
pytz>=2021.1
attrs>=21.0.0
packaging>=21.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Análisis interactivo (Jupyter)
ipython>=7.20.0
ipykernel>=6.0.0
jupyter>=1.0.0

# Proyecto local (instalar paquete del repo)
.
